requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
python-multipart==0.0.20
httpx==0.28.1
python-dotenv
PyPDF2
Pillow
openai
